Data.Par.Range

Description

A mechanism to describe iteration spaces supporting parallel execution.

Constructing iteration spaces

Beware that ranges include information about whether or not sequential bottoming-out is permitted. If you are executing parallel loops with inter-iteration dependencies, be careful to use fullpar.

data Range

An iteration space expressed as an inclusive range of integers, i.e. `InclusiveRange 1 3` includes both `1` and `3`.

An iteration space also includes information that determines splitting behavior (e.g. threshold to bottom out to sequential execution).

Constructors

InclusiveRange

Fields

startInd :: !Int

Start, inclusive

endInd :: !Int

End, inclusive

seqThresh :: !Int

For ranges less than or equal to this, switch to sequential execution.

range :: Int -> Int -> Range

A simple shorthand for ranges from n to `m-1` (inclusive,exclusive).

Note that this function, as well as irange and zrange, by default produce "auto-sequentializing" iteration spaces that choose a threshold for bottoming out to sequential execution based on the size of the range and the number of processors.

irange :: Int -> Int -> Range

A simple shorthand for inclusive ranges from n to m.

zrange :: Int -> Range

A simple shorthand for ranges from `0` to `n-1`

fullpar :: Range -> Range

Tweak an iteration range to exploit all parallelism. That is, never bottom-out to sequential loops.

Combined MapReduce operations on ranges

pmapReduce

Arguments

:: (NFData a, ParFuture p, FutContents p a)

=> Range
iteration range over which to calculate

-> (Int -> p a)
compute one result

-> (a -> a -> p a)
combine two results

-> a
initial result

-> p a

Computes a binary map/reduce over a finite range. The range is recursively split into two, the result for each half is computed in parallel, and then the two results are combined. The thresholding behavior for the range is obeyed, and a sequential loop is used if the splitting bottoms out.

For example, the following is a parallel implementation of

  foldl (+) 0 (map (^2) [1..10^6])

 parMapReduce (irange 1 (10^6))
        (\x -> return (x^2))
        (\x y -> return (x+y))
        0

If an automatic threshold is being used in the underlying ranges, answers may vary on different machines. But note that this function does NOT require a commutative combining function. Commutativity, however, is not required. This will always combine lower iteration results on the left and higher on the right.

pmapReduce_

Arguments

:: (ParFuture p, FutContents p a)

=> Range
iteration range over which to calculate

-> (Int -> p a)
compute one result

-> (a -> a -> p a)
combine two results

-> a
initial result

-> p a

A version of pmapReduce that is only weak-head-normal-form (WHNF) strict in the folded accumulators.

Misc

toList :: Range -> [Int]

Enumerate the elements in a Range.