Data.Par.Splittable

Description

A collection of useful parallel combinators based on top of a Par monad together with Data.Splittable.

Efficient, balanced parallel traversals for splittable structures

These operations require an instance of Split, but in return they can perform more balanced traversals that are more tolerant of fine-granularity.

pmapReduce

Arguments

:: forall c e m a t . (Split c, Generator c, ParFuture m, FutContents m a, NFData a)

=> c
element generator to consume

-> (ElemOf c -> m a)
compute one result

-> (a -> a -> m a)
combine two results

-> a
initial accumulator value

-> m a

Computes a binary map/reduce over a data source. The input is recursively split into two, the result for each half is computed in parallel, and then the two results are combined. A sequential loop is used if the splitting bottoms out.

For example, the following is a parallel implementation of

  foldl (+) 0 (map (^2) [1..10^6])

 import Data.Par.Range (irange)
  ...
 parMapReduce (irange 1 (10^6))
        (\x -> return (x^2))
        (\x y -> return (x+y))
        0

If an automatic threshold is being used in the underlying ranges, answers may vary on different machines if the reduce function is not associative. But even then, platform portability does NOT require a commutative reduce function. This combinator will always fold earlier results on the left and later on the right.

pmapReduce_

Arguments

:: forall c e m a t . (Split c, Generator c, ParFuture m, FutContents m a)

=> c
element generator to consume

-> (ElemOf c -> m a)
compute one result

-> (a -> a -> m a)
combine two results

-> a
initial accumulator value

-> m a

A version of pmapReduce that is only weak-head-normal-form (WHNF) strict in the folded accumulators.

pmapReduceWith_

Arguments

:: forall c e m a t . (Generator c, ParFuture m, FutContents m a)

=> (c -> [c])
splitting function.

-> c
element generator to consume

-> (ElemOf c -> m a)
compute one result

-> (a -> a -> m a)
combine two results

-> a
initial accumulator value

-> m a

A version of pmapReduce_ that uses a custom splitting function.

pforEach

Arguments

:: (Split c, Generator c, ParFuture m, FutContents m ())

=> c
element generator to consume

-> (ElemOf c -> m ())
compute one result

-> m ()

Execute a side-effect for each element generated. Use the Split instance to determine the degree of parallelism (granularity).

This is SYNCHRONOUS; that is, it does not return until all of the actions are executed.

asyncForEach

Arguments

:: (Split c, Generator c, ParFuture m, FutContents m ())

=> c
element generator to consume

-> (ElemOf c -> m ())
compute one result

-> m ()

Non-blocking version of pforEach.

The underlying, general version of which the above are specializations

mkMapReduce

Arguments

:: forall c e m a t . (ParFuture m, FutContents m a)

=> (c -> [c])
splitting function

-> ((a -> e -> m a) -> a -> c -> m a)
sequential fold function

-> (m a -> m (Future m a))
spawn function

-> c
element generator to consume

-> (e -> m a)
compute one result

-> (a -> a -> m a)
combine two results

-> a
initial accumulator value

-> m a

Make a parallel map-reduce function given a custom function for spawning work.