Data.Par.Traversable

Description

Parallel combinators based on top of a Par monad. Specifically, this module provides higher order functions for operating on Traversable data structures in parallel.

Naive parallel maps on traversable structures.

Because these operations assume only Traversable, the best they can do is to fork one parallel computation per element.

parMap :: (Traversable t, NFData b, ParFuture p, FutContents p b) => (a -> b) -> t a -> p (t b)

Applies the given function to each element of a data structure in parallel (fully evaluating the results), and returns a new data structure containing the results.

 parMap f xs = mapM (spawnP . f) xs >>= mapM get

pmap is commonly used for lists, where it has this specialised type:

 parMap :: NFData b => (a -> b) -> [a] -> Par [b]

But note that for efficient parallelism you want balanced task trees, not one at a time parallel tasks. Thus look at pmapReduce and friends.

ptraverse :: (Traversable t, NFData b, ParFuture p, FutContents p b) => (a -> p b) -> t a -> p (t b)

Like parMap, but the function is a Par monad operation.

 ptraverse f xs = mapM (spawn . f) xs >>= mapM get

Variants that evaluate only down to weak-head-normal-form

parMap_ :: (Traversable t, ParFuture p, FutContents p b) => (a -> b) -> t a -> p (t b)

A variant of ParMap that only evaluates to weak-head-normal-form.

ptraverse_ :: (Traversable t, ParFuture p, FutContents p b) => (a -> p b) -> t a -> p (t b)

A variant that only evaluates to weak-head-normal-form.